Added lwjgl 2.9.1
Added lwjgl to the repository and set it up in the project properties.
Please make sure it is set up by going to File/Project Properties/ in
Netbeans go to the libraries tab and make sure lwjgl is added. If not
then click "Add JAR/folder", go to the location of the repository on
your computer and add the lwjgl-2.9.1.zip file to the libraries.