Added "last position"
added last position and angle so there is only an update to the log if
the camera actually moved. this will significantly decrease the size of
any logs we create as well as just be much easier to read as our last
movement won't be incredibly far behind in the log.